requirements.txt
absl-py==0.11.0
aiohttp==3.7.4.post0
apturl==0.5.2
astunparse==1.6.3
async-timeout==3.0.1
attrs==19.3.0
awscli==1.20.7
bcrypt==3.1.7
beautifulsoup4==4.9.3
blinker==1.4
botocore==1.21.7
Brlapi==0.7.0
CacheControl==0.12.6
cached-property==1.5.1
cachetools==4.0.0
capstone==4.0.2
certifi==2019.11.28
cffi==1.14.4
chardet==3.0.4
click==8.0.1
colorama==0.4.3
colored-traceback==0.3.0
command-not-found==0.3
crypto==1.4.1
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
defer==1.0.6
discord.py==1.7.3
distro==1.4.0
distro-info===0.23ubuntu1
dnspython==2.1.0
docker==4.1.0
docker-compose==1.25.0
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
firebase-admin==4.5.1
Flask==2.0.1
flatbuffers==1.12
future==0.18.2
gast==0.3.3
gmpy2==2.0.8
google-api-core==1.25.0
google-api-python-client==1.7.11
google-auth==1.24.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.2
google-cloud-core==1.5.0
google-cloud-firestore==2.0.2
google-cloud-speech==2.0.1
google-cloud-storage==1.35.1
google-crc32c==1.1.2
google-pasta==0.2.0
google-resumable-media==1.2.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
h5py==2.10.0
html5lib==1.1
httplib2==0.14.0
idna==2.8
img2pdf==0.3.3
impacket==0.9.23
importlib-metadata==1.5.0
intervaltree==3.1.0
itsdangerous==2.0.1
jikuai==0.0.8
Jinja2==3.0.1
jmespath==0.10.0
joblib==1.0.1
Js2Py==0.70
jsonschema==3.2.0
Keras-Preprocessing==1.1.2
keyring==18.0.1
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
ldap3==2.9.1
ldapdomaindump==0.9.3
libcst==0.3.16
lockfile==0.12.2
louis==3.12.0
lxml==4.6.2
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.3.3
MarkupSafe==2.0.1
matplotlib==3.3.3
mechanize==0.4.5
monotonic==1.5
more-itertools==4.2.0
msgpack==1.0.2
multidict==5.1.0
mypy-extensions==0.4.3
Naked==0.1.31
netifaces==0.10.4
numpy==1.19.2
oauth2client==4.1.2
oauthlib==3.1.0
olefile==0.46
opencv-contrib-python==4.5.1.48
opencv-python==4.5.1.48
opt-einsum==3.3.0
packaging==20.8
pandas==1.2.1
pandas-datareader==0.9.0
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.0.0
Pillow-PIL==0.1.dev0
pipwin==0.5.0
plumbum==1.7.0
powerline-status==2.7
proto-plus==1.13.0
protobuf==3.14.0
psutil==5.5.1
pwntools==4.5.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
PyAudio==0.2.11
PyBluez==0.23
pycairo==1.16.2
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.9.7
pycryptodomex==3.10.1
pycups==1.9.73
PyDrive==1.3.1
pydub==0.24.1
pyelftools==0.27
Pygments==2.3.1
PyGObject==3.36.0
pyjsparser==2.7.1
PyJWT==1.7.1
pymacaroons==0.13.0
PyMySQL==1.0.2
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
PyPrind==2.11.2
pyRFC3339==1.1
pyrsistent==0.15.5
pyserial==3.5
pySmartDL==1.3.4
PySocks==1.7.1
python-apt==2.0.0+ubuntu0.20.4.5
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-dotenv==0.19.0
pytz==2019.3
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
ROPGadget==6.5
rpyc==5.0.1
rsa==4.0
s3transfer==0.5.0
sage==9.0
scikit-learn==0.24.1
scipy==1.6.0
SecretStorage==2.3.1
shellescape==3.8.1
simplejson==3.16.0
six==1.15.0
sklearn==0.0
sortedcontainers==2.3.0
soupsieve==2.1
SpeechRecognition==3.8.1
speedtest-cli==2.1.2
SQLAlchemy==1.3.12
systemd-python==234
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
texttable==1.6.2
threadpoolctl==2.1.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
tzlocal==2.1
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
unicorn==1.0.2rc3
uritemplate==0.6
urllib3==1.25.8
usb-creator==0.3.7
wadllib==1.3.3
webencodings==0.5.1
websocket-client==0.53.0
Werkzeug==2.0.1
wrapt==1.12.1
xkit==0.0.0
xlrd==2.0.1
XlsxWriter==1.3.7
xlwt==1.3.0
yarl==1.6.3
youtube-dl==2020.3.24
zipp==1.0.0
